Move search to its own package.
